introduced retry functionality for calculated service metrics
Calculated Service Metrics can depend on Request Attributes. If said Request Attribute has been applied immediately before the service metric, creation of that metric is prone to fail, because the Dynatrace Cluster is not fully aware of its existence yet.
If the REST API responds back with an error stating that the Request Attribute with the given name is not known, the REST client will retry the POST request.