feat: Define version and use `ldflags` to set version during GoReleaser build
Define `version.Current`, `version.OperatingSystem` and `version.CPUArchitecture` as variables for the current version, OS and CPU architecture.
These must be variables so that they can be set.
The GoReleaser configuration is updated to set these during a build.
Also, `GetUserAgent()` is added to get a standard user-agent based on these fields.