test: add E2E tests
This adds the E2E tests for the pipeline groups. The examples are mostly identical.
It's two pipelines, an empty one and one with davis/metricExtractions/securityContext defined, depending on whether it's enabled on that resource.
One pipeline group that creates a group with the stages of a pipeline (the second-mentioned one) and assigns it to another pipeline (the first-mentioned one)