feat: allow the use of feature flags
This allows us to make use of feature flags. With this, features/resources can be enabled/disabled.
Some refactoring had to be done because even if `AllResources` is set to the correct values according to the envs, this can't be tested in tests, because at the time we are setting the env in the test, the value is already set