feat(aws-connection): Adds retry on role ARN update
This is necessary because when the IAM role is created on AWS, it takes some time until it comes into effect. If immediately after creating the IAM role the role ARN is added to the AWS connection, the HAS will fail when trying to assume the role. Therefore, we need to retry adding the role ARN