fix: Active Gate URL parsing
For URLs like `myhost:9999/e/my-env-id` the classic requests always attached another `/e/my-env-id`, which of course broke the requests.
Instead of conditionally removing a URL prefix, the initial path is not added in the first place. The API client that's in use already contains this information. That's why the path was attached a second time.