fix: Check error first before processing response when using `settings20.Client`
While the current code doesn't panic as an empty Response is returned for errors with `StatusCode` set to `0`,  the actual error is lost. This change fixes that by checking the error first.

The `settings20.Client` is only used in these four services.

This change makes the code more Go idiomatic.