fix: `CreateClassicOAuthBasedClient()` returns `nil` if client ID or secret are not set
By returning `nil` in this situation, the function no longer returns a client that has an API token set. This is OK because in all instances where this client is used, an explicit client with API token set is created and used first.

The function also no longer returns an error as this was always dropped.